Add .gitignore entries for Xcode-related files
This commit introduces new entries to the `.gitignore` file to exclude Xcode-specific files and folders from version control. This ensures that only relevant project files are tracked, simplifying the repository and reducing clutter.

Specifically, it ignores `staged_diff.txt` files, which are Xcode-specific files used for staging changes during development, and the `.vscode` folder, which contains user-specific Visual Studio Code settings that are not relevant for version control.